+++ src/ts/BezierPath.ts
@@ -1,7 +1,7 @@
 /**
  * @classdesc A refactored BezierPath class.
  *
- * @require Vertex, CubicBezierCurve
+ * @require Bounds, Vertex, CubicBezierCurve, XYCoords, SVGSerializable
  * 
  * @author Ikaros Kappler
  * @date 2013-08-19
@@ -18,12 +18,15 @@
  * @modified 2020-02-06 Added function locateCurveByEndPoint( Vertex ).
  * @modified 2020-02-11 Added 'return this' to the scale(Vertex,number) and to the translate(Vertex) function.
  * @modified 2020-03-24 Ported this class from vanilla-JS to Typescript.
- * @version 2.1.2
+ * @modified 2020-06-03 Made the private helper function _locateUIndex to a private function.
+ * @modified 2020-06-03 Added the getBounds() function.
+ * @version 2.2.0
  *
  * @file BezierPath
  * @public
  **/
 
+import { Bounds } from "./Bounds";
 import { CubicBezierCurve } from "./CubicBezierCurve";
 import { Vertex } from "./Vertex";
 import { XYCoords, SVGSerializable} from "./interfaces";
@@ -776,7 +779,7 @@ export class BezierPath implements SVGSerializable {
      * - {number} uPart - the absolute curve length sum (length from the beginning to u, should equal u itself).
      * - {number} uBefore - the absolute curve length for all segments _before_ the matched curve (usually uBefore <= uPart).
      **/
-    static _locateUIndex( path:BezierPath, u:number ) : { i:number, uPart:number, uBefore:number } {
+    private static _locateUIndex( path:BezierPath, u:number ) : { i:number, uPart:number, uBefore:number } {
 	var i : number = 0;
 	var uTemp : number = 0.0;
 	var uBefore : number = 0.0;
@@ -940,15 +943,15 @@ export class BezierPath implements SVGSerializable {
 
 	var mainCurve : CubicBezierCurve = this.getCurveAt( curveIndex );
 	var neighbourCurve : CubicBezierCurve = this.getCurveAt( curveIndex-1<0 ? this.getCurveCount()+(curveIndex-1) : curveIndex-1 );
-	/* return ? */ this.adjustNeighbourControlPoint( mainCurve,
-						 neighbourCurve,
-						 mainCurve.getStartPoint(),            // the reference point
-						 mainCurve.getStartControlPoint(),     // the dragged control point
-						 neighbourCurve.getEndPoint(),         // the neighbour's point
-						 neighbourCurve.getEndControlPoint(),  // the neighbour's control point to adjust
-						 obtainHandleLength,
-						 updateArcLengths
-					       );
+	BezierPath.adjustNeighbourControlPoint( mainCurve,
+						neighbourCurve,
+						mainCurve.getStartPoint(),            // the reference point
+						mainCurve.getStartControlPoint(),     // the dragged control point
+						neighbourCurve.getEndPoint(),         // the neighbour's point
+						neighbourCurve.getEndControlPoint(),  // the neighbour's control point to adjust
+						obtainHandleLength,
+						updateArcLengths
+					      );
     };
 
 
@@ -971,15 +974,15 @@ export class BezierPath implements SVGSerializable {
 
 	var mainCurve : CubicBezierCurve      = this.getCurveAt( curveIndex );
 	var neighbourCurve : CubicBezierCurve = this.getCurveAt( (curveIndex+1)%this.getCurveCount() );
-	return this.adjustNeighbourControlPoint( mainCurve,
-						 neighbourCurve,
-						 mainCurve.getEndPoint(),                // the reference point
-						 mainCurve.getEndControlPoint(),         // the dragged control point
-						 neighbourCurve.getStartPoint(),         // the neighbour's point
-						 neighbourCurve.getStartControlPoint(),  // the neighbour's control point to adjust
-						 obtainHandleLength,
-						 updateArcLengths
-					       );
+	/* return */ BezierPath.adjustNeighbourControlPoint( mainCurve,
+							     neighbourCurve,
+							     mainCurve.getEndPoint(),                // the reference point
+							     mainCurve.getEndControlPoint(),         // the dragged control point
+							     neighbourCurve.getStartPoint(),         // the neighbour's point
+							     neighbourCurve.getStartControlPoint(),  // the neighbour's control point to adjust
+							     obtainHandleLength,
+							     updateArcLengths
+							   );
     };
 
     /**
@@ -1000,15 +1003,15 @@ export class BezierPath implements SVGSerializable {
      * @return {void}
      **/
     // !!! TODO: SHOULDNT THIS BE A STATIC FUNCTION ???
-    adjustNeighbourControlPoint( mainCurve:CubicBezierCurve,
-				 neighbourCurve:CubicBezierCurve,
-				 mainPoint:Vertex,
-				 mainControlPoint:Vertex,
-				 neighbourPoint:Vertex,
-				 neighbourControlPoint:Vertex,
-				 obtainHandleLengths:boolean, 
-				 updateArcLengths:boolean
-			       ) : void {
+    private static adjustNeighbourControlPoint( mainCurve:CubicBezierCurve,
+						neighbourCurve:CubicBezierCurve,
+						mainPoint:Vertex,
+						mainControlPoint:Vertex,
+						neighbourPoint:Vertex,
+						neighbourControlPoint:Vertex,
+						obtainHandleLengths:boolean, 
+						updateArcLengths:boolean
+					      ) : void {
 
 	// Calculate start handle length
 	var mainHandleBounds : Vertex       = new Vertex( mainControlPoint.x - mainPoint.x,
@@ -1029,11 +1032,34 @@ export class BezierPath implements SVGSerializable {
 				       neighbourPoint.y - mainHandleBounds.y * (neighbourHandleLength/mainHandleLength)
 				     );
 	} else {
-	    neighbourControlPoint.set( neighbourPoint.x - mainHandleBounds.x, // * (neighbourHandleLength/mainHandleLength),
-				       neighbourPoint.y - mainHandleBounds.y // * (neighbourHandleLength/mainHandleLength)
+	    neighbourControlPoint.set( neighbourPoint.x - mainHandleBounds.x,
+				       neighbourPoint.y - mainHandleBounds.y 
 				     );
 	}
 	neighbourCurve.updateArcLengths();
+    };
+
+
+    /**
+     * Get the bounds of this Bézier path.
+     *
+     * Note the the curves' underlyung segment buffers are used to determine the bounds. The more
+     * elements the segment buffers have, the more precise the returned bounds will be.
+     *
+     * @return {Bounds} The bounds of this Bézier path.
+     **/
+    getBounds() : Bounds {
+	const min : Vertex = new Vertex( Number.POSITIVE_INFINITY, Number.POSITIVE_INFINITY );
+	const max : Vertex = new Vertex( Number.NEGATIVE_INFINITY, Number.NEGATIVE_INFINITY );
+	var b : Bounds;
+	for( var i = 0; i < this.bezierCurves.length; i++ ) {
+	    b = this.bezierCurves[i].getBounds();
+	    min.x = Math.min( min.x, b.min.x );
+	    min.y = Math.min( min.y, b.min.y );
+	    max.x = Math.max( max.x, b.max.x );
+	    max.y = Math.max( max.y, b.max.y );
+	}
+	return new Bounds( min, max );
     };