+++ src/ts/draw.ts
@@ -166,7 +166,7 @@ export class drawutils {
      * @instance
      * @memberof drawutils
      **/
-    image( image:HTMLImageElement, position:Vertex, size:Vertex ) {
+    image( image:HTMLImageElement, position:Vertex, size:Vertex ) : void {
 	this.ctx.save();
 	// Note that there is a Safari bug with the 3 or 5 params variant.
 	// Only the 9-param varaint works.
@@ -179,6 +179,29 @@ export class drawutils {
 			    size.x*this.scale.x,
 			    size.y*this.scale.y );
 	this.ctx.restore();	
+    };
+
+
+    /**
+     * Draw a rectangle.
+     *
+     * @param {Vertex} position - The upper left corner of the rectangle.
+     * @param {number} width - The width of the rectangle.
+     * @param {number} height - The height of the rectangle.
+     * @param {string} color - The color to use.
+     * @param {number=1} lineWidth - (optional) The line with to use (default is 1).
+     **/
+    rect( position:Vertex, width:number, height:number, color:string, lineWidth?:number ) : void {
+	this.ctx.save();
+	this.ctx.beginPath();
+	this.ctx.moveTo( this.offset.x+position.x*this.scale.x, this.offset.y+position.y*this.scale.y );
+	this.ctx.lineTo( this.offset.x+(position.x+width)*this.scale.x, this.offset.y+position.y*this.scale.y );
+	this.ctx.lineTo( this.offset.x+(position.x+width)*this.scale.x, this.offset.y+(position.y+height)*this.scale.y );
+	this.ctx.lineTo( this.offset.x+position.x*this.scale.x, this.offset.y+(position.y+height)*this.scale.y );
+	this.ctx.closePath();
+	this.ctx.lineWidth = lineWidth || 1;
+	this._fillOrDraw( color );
+	this.ctx.restore();
     };